+++ src/main/java/ch/csbe/productmanager/resources/product/ProductRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Repository für den Zugriff auf die Produkte in der Datenbank.
  */
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
 }
